update intégration sprint 2

+++ back_matagri_projet_3/src/main/java/mg/inclusiv/cdan8/projet3/Entities/Materiels.java
@@ -30,6 +30,6 @@ public class Materiels {
     // @ManyToOne
     // @JoinColumn(name = "id_user")
     // private Users user;
-    private Long id_user;
+    private Integer id_user;
     
 }
